requirements.txt
av==9.2.0
numpy==1.21.5
pandas==1.4.2
Pillow==9.3.0
streamlit==1.15.2
streamlit_webrtc==0.44.0
torch==1.13.0
yolov5==6.2.3
opencv-python-headless==4.6.0.66
XlsxWriter==3.0.8
